+++ src/shared/settings/index.js
@@ -1,4 +1,5 @@
 export default {
   timeout: 5000,
   showCloseButton: false,
+  enableClickAwayListener: false,
 };