+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.0
 - [new] Remove 'palette' wrapper object from theme object.
 - [new] Attach adornment icons to theme object.
+- [new] Optional click away listener setting.
 
 ## 0.3.2
 - [fix] Update global theme through context.